refactor: rename cart_pole_cost module
This commit renames the `cart_pole_cost` module to `cartpole_cost` for
consitency with the gymnasium package.
