+++ simzoo/envs/classic_control/__init__.py
@@ -6,7 +6,7 @@ import sys
 # NOTE: Makes sure that it works both in the Simzoo stand-alone package and the
 # name_space package (blc)
 if "simzoo" in sys.modules:
-    from simzoo.envs.classic_control.cart_pole_cost.cart_pole_cost import CartPoleCost
+    from simzoo.envs.classic_control.cartpole_cost.cartpole_cost import CartPoleCost
     from simzoo.envs.classic_control.ex3_ekf.ex3_ekf import Ex3EKF
 elif importlib.util.find_spec("simzoo") is not None:
     Ex3EKF = getattr(
@@ -15,7 +15,7 @@ elif importlib.util.find_spec("simzoo") is not None:
     )
     CartPoleCost = getattr(
         importlib.import_module(
-            "simzoo.envs.classic_control.cart_pole_cost.cart_pole_cost"
+            "simzoo.envs.classic_control.cartpole_cost.cartpole_cost"
         ),
         "CartPoleCost",
     )
@@ -28,7 +28,7 @@ else:
     )
     CartPoleCost = getattr(
         importlib.import_module(
-            "bayesian_learning_control.simzoo.simzoo.envs.classic_control.cart_pole_cost.cart_pole_cost"  # noqa: E501
+            "bayesian_learning_control.simzoo.simzoo.envs.classic_control.cartpole_cost.cartpole_cost"  # noqa: E501
         ),
         "CartPoleCost",
     )
